Made required changes
